chore: Extract slider component
