+++ app/src/main/java/io/github/zyrouge/symphony/ui/view/settings/SliderTile.kt
@@ -1,25 +1,11 @@
 package io.github.zyrouge.symphony.ui.view.settings
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.gestures.detectHorizontalDragGestures
-import androidx.compose.foundation.gestures.detectTapGestures
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxWithConstraints
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Card
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ListItem
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ProvideTextStyle
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
@@ -29,11 +15,10 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.dp
 import io.github.zyrouge.symphony.ui.components.ScaffoldDialog
+import io.github.zyrouge.symphony.ui.components.Slider
 import io.github.zyrouge.symphony.ui.helpers.ViewContext
-import io.github.zyrouge.symphony.utils.RangeUtils
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -66,10 +51,6 @@ fun SettingsSliderTile(
 
     if (isOpen) {
         var value by remember { mutableFloatStateOf(initialValue) }
-        var ratio by remember {
-            mutableFloatStateOf(RangeUtils.calculateRatioFromValue(value, range))
-        }
-        var pointerOffsetX = 0f
 
         ScaffoldDialog(
             onDismissRequest = {
@@ -77,79 +58,17 @@ fun SettingsSliderTile(
             },
             title = title,
             content = {
-                Column(
+                Slider(
+                    value = value,
+                    onChange = { nValue ->
+                        value = onValue(nValue)
+                    },
+                    range = range,
+                    label = label,
                     modifier = Modifier
                         .padding(top = 16.dp)
                         .verticalScroll(rememberScrollState())
-                ) {
-                    BoxWithConstraints(modifier = Modifier.padding(20.dp, 0.dp)) {
-                        val height = 12.dp
-                        val shape = RoundedCornerShape(height.div(2))
-
-                        Box(
-                            modifier = Modifier
-                                .background(
-                                    MaterialTheme.colorScheme.surfaceVariant,
-                                    shape,
-                                )
-                                .fillMaxWidth()
-                                .height(height)
-                                .pointerInput(Unit) {
-                                    detectTapGestures(
-                                        onTap = { offset ->
-                                            pointerOffsetX = offset.x
-                                            val widthPx = maxWidth.toPx()
-                                            val nRatio = (pointerOffsetX / widthPx).coerceIn(0f..1f)
-                                            val nValue =
-                                                RangeUtils.calculateValueFromRatio(nRatio, range)
-                                            value = onValue(nValue)
-                                            ratio = RangeUtils.calculateRatioFromValue(value, range)
-                                        }
-                                    )
-                                }
-                                .pointerInput(Unit) {
-                                    detectHorizontalDragGestures(
-                                        onDragStart = { offset ->
-                                            pointerOffsetX = offset.x
-                                        },
-                                        onHorizontalDrag = { pointer, dragAmount ->
-                                            pointer.consume()
-                                            pointerOffsetX += dragAmount
-                                            val widthPx = maxWidth.toPx()
-                                            val nRatio = (pointerOffsetX / widthPx).coerceIn(0f..1f)
-                                            val nValue =
-                                                RangeUtils.calculateValueFromRatio(nRatio, range)
-                                            value = onValue(nValue)
-                                            ratio = RangeUtils.calculateRatioFromValue(value, range)
-                                        },
-                                    )
-                                }
-                        ) {
-                            Box(
-                                modifier = Modifier
-                                    .background(
-                                        MaterialTheme.colorScheme.primary,
-                                        shape,
-                                    )
-                                    .fillMaxWidth(ratio)
-                                    .height(height)
-                            )
-                        }
-                    }
-                    Spacer(modifier = Modifier.height(8.dp))
-                    ProvideTextStyle(MaterialTheme.typography.labelMedium) {
-                        Row(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(20.dp, 0.dp),
-                            horizontalArrangement = Arrangement.SpaceBetween,
-                        ) {
-                            label(range.start)
-                            label(value)
-                            label(range.endInclusive)
-                        }
-                    }
-                }
+                )
             },
             actions = {
                 onReset?.let {
